requirements.txt
google-genai>=0.1.0
openai>=1.1.0
playwright>=1.40.0
python-dotenv>=1.0.0
requests>=2.31.0
rembg[gpu]>=2.0.50
Pillow>=10.2.0
